agents/chainlit-adk-101/pyproject.toml
[project]
name = "chainlit101"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "google-adk>=0.3.0",
    "python-dotenv>=1.1.0",
]
